Update
Agent no longer teleports for A* traversing, but can move rather quickly for my liking (I need to check frame rates for this to see if it's misbehaving or not). The Step function in agent is getting a little hefty. Next update will likely clean that up and start on metacognition